controllers/version/metadata: Move `update_version()` to dedicated module
